requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiomysql==0.2.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
email-validator==2.1.0

# ML/AI Dependencies
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
spacy==3.7.2
nltk==3.8.1

# Additional utilities
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6
